+++ .idea/runConfigurations/Launcher.xml
@@ -1,7 +1,8 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Launcher" type="Application" factoryName="Application" nameIsGenerated="true">
-    <option name="MAIN_CLASS_NAME" value="com.leti.summer_practice.Launcher" />
-    <module name="Summer_Practice" />
+  <configuration default="false" name="Launcher" type="Application" factoryName="Application">
+    <option name="MAIN_CLASS_NAME" value="com.borviz.boruvkavisualizer.Launcher" />
+    <module name="BoruvkaVisualizer" />
+    <option name="WORKING_DIRECTORY" value="$MODULE_WORKING_DIR$" />
     <extension name="coverage">
       <pattern>
         <option name="PATTERN" value="com.leti.summer_practice.*" />
