+++ src/main/java/com/borviz/boruvkavisualizer/Launcher.java
@@ -1,0 +1,10 @@
+package com.borviz.boruvkavisualizer;
+
+import com.borviz.boruvkavisualizer.gui.prog.BorVizApplication;
+
+public class Launcher {
+
+    public static void main(String[] args) {
+        BorVizApplication.launchBorVizApplication(args);
+    }
+}
