+++ src/main/java/com/borviz/boruvkavisualizer/gui/prog/BorVizApplication.java
@@ -1,17 +1,17 @@
-package com.leti.summer_practice.gui.prog;
+package com.borviz.boruvkavisualizer.gui.prog;
 
-import com.leti.summer_practice.R;
+import com.borviz.boruvkavisualizer.R;
 import javafx.application.Application;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
-public class SummerPracticeApplication extends Application {
+public class BorVizApplication extends Application {
 
-    private static SummerPracticeApplication application;
+    private static BorVizApplication application;
 
     private Stage primaryStage;
 
-    public static void launchSummerPracticeApplication(String... args) {
+    public static void launchBorVizApplication(String... args) {
         launch(args);
     }
 
@@ -20,11 +20,11 @@ public class SummerPracticeApplication extends Application {
         application = this;
         setPrimaryStage(primaryStage);
         primaryStage.setTitle(R.string("AppName"));
-        primaryStage.setScene(new Scene(R.loadFXML("layouts/summer-practice.fxml"), 800, 600));
+        primaryStage.setScene(new Scene(R.loadFXML("layouts/boruvka-visualizer.fxml"), 800, 600));
         primaryStage.show();
     }
 
-    public static SummerPracticeApplication getApplication() {
+    public static BorVizApplication getApplication() {
         return application;
     }
