--- src/main/java/com/leti/summer_practice/Launcher.java
@@ -1,10 +1,0 @@
-package com.leti.summer_practice;
-
-import com.leti.summer_practice.gui.prog.SummerPracticeApplication;
-
-public class Launcher {
-
-    public static void main(String[] args) {
-        SummerPracticeApplication.launchSummerPracticeApplication(args);
-    }
-}
