+++ src/main/java/module-info.java
@@ -1,4 +1,4 @@
-module com.leti.summer_practice {
+module com.borviz.boruvkavisualizer {
 
     requires javafx.controls;
     requires javafx.fxml;
@@ -10,10 +10,10 @@ module com.leti.summer_practice {
     requires org.kordamp.bootstrapfx.core;
     requires eu.hansolo.tilesfx;
 
-    exports com.leti.summer_practice.logic;
+    exports com.borviz.boruvkavisualizer.logic;
 
-    opens com.leti.summer_practice.gui.lib to javafx.fxml;
-    exports com.leti.summer_practice.gui.lib;
-    opens com.leti.summer_practice.gui.prog to javafx.fxml;
-    exports com.leti.summer_practice.gui.prog;
+    opens com.borviz.boruvkavisualizer.gui.lib to javafx.fxml;
+    exports com.borviz.boruvkavisualizer.gui.lib;
+    opens com.borviz.boruvkavisualizer.gui.prog to javafx.fxml;
+    exports com.borviz.boruvkavisualizer.gui.prog;
 }
