+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,2 +1,2 @@
 Manifest-Version: 1.0
-Main-Class: com.leti.summer_practice.Launcher
+Main-Class: com.borviz.boruvkavisualizer.Launcher
