+++ src/main/resources/com/borviz/boruvkavisualizer/strings/strings.properties
