+++ src/main/resources/com/borviz/boruvkavisualizer/strings/strings_en.properties
@@ -1,4 +1,4 @@
-AppName=Summer Practice
+AppName=Boruvka Visualizer
 draw_mode=Draw mode
 drag_mode=Drag mode
 load_from_file=Load from file
