+++ src/main/resources/com/borviz/boruvkavisualizer/strings/strings_ru.properties
@@ -1,4 +1,4 @@
-AppName=Летняя практика
+AppName=Визуализатор Борувки
 draw_mode=Режим рисования
 drag_mode=Режим перетаскивания
 load_from_file=Загрузить из файла
